ramips: mt76x8: add support for Cudy LT400E v1 (WiP)
This commit adds support for the Cudy LT400E v1, an indoor 4G/LTE WiFi
router with 4x10/100 Mbps Ethernet ports.

Hardware:
- SoC: MediaTek MT7628NN
- Flash: 8 MiB SPI
- RAM: 64 MiB
- Ethernet: 4x10/100 Mbps
- WiFi: 2.4 GHz (embedded MT7628, 802.11n)
- WWAN: Quectel EC200A-EL 4G/LTE Cat 4 modem
- Buttons: 1 Reset button, 1 WPS button
- LEDs: 10x (blue)
- Power: 12 V DV barrel jack

Acknowledgments:
The code to support this device is largely based on the sources kindly
provided by Cudy in compliance with the GPL license. Thanks for
supporting the OpenWrt community.

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@guifi.net>